PubSub Enhancement - onNewListeners
Adds an additional method to allow a Subscriber to be informed when a new Listener is registered to a specific event. 
Allowing the Subscriber to e.g. send a welcome/initial event to make sure all listeners start with the same values.

Use-case: Style Distributor, which shares one page style to all components on that page.

Code example on how to apply:
`onNewListeners( STYLE_EVENT_NAME, this.publishStyles, this );`

https://github.com/foxcreation/LWC-Style-Distributor